+++ news-bundle/src/Resources/contao/modules/ModuleNewsArchive.php
@@ -142,8 +142,8 @@ class ModuleNewsArchive extends \ModuleNews
 			}
 			elseif ($this->news_jumpToCurrent == 'all_items')
 			{
-				$intBegin = 0;
-				$intEnd = time();
+				$intBegin = 0; // 1970-01-01 00:00:00
+				$intEnd = 2145913200; // 2038-01-01 00:00:00
 			}
 		}
 		catch (\OutOfBoundsException $e)