+++ core-bundle/src/Resources/contao/drivers/DC_Folder.php
@@ -2769,15 +2769,14 @@ class DC_Folder extends \DataContainer implements \listable, \editable
 		// Store search value in the current session
 		if (\Input::post('FORM_SUBMIT') == 'tl_filters')
 		{
-			$strField = \Input::post('tl_field', true);
 			$strKeyword = ltrim(\Input::postRaw('tl_value'), '*');
 
 			// Make sure the regular expression is valid
-			if ($strKeyword != '')
+			if ($strKeyword)
 			{
 				try
 				{
-					$this->Database->prepare("SELECT * FROM " . $this->strTable . " WHERE " . $strField . " REGEXP ?")
+					$this->Database->prepare("SELECT * FROM " . $this->strTable . " WHERE name REGEXP ?")
 								   ->limit(1)
 								   ->execute($strKeyword);
 				}
@@ -2787,7 +2786,6 @@ class DC_Folder extends \DataContainer implements \listable, \editable
 				}
 			}
 
-			$session['search'][$this->strTable]['field'] = $strField;
 			$session['search'][$this->strTable]['value'] = $strKeyword;
 
 			$objSessionBag->replace($session);